requirements.txt
# PDF処理
PyMuPDF>=1.23.0

# Claude API
anthropic>=0.18.0

# 環境変数管理
python-dotenv>=1.0.0

# テスト
pytest>=8.0.0
pytest-cov>=4.1.0
